requirements.txt
aiofiles
Pillow
humanize
hachoir
motor
pyrogram==2.0.106
TgCrypto
dnspython
aiohttp 
pyromod==1.5
Telethon==1.*
